requirements.txt
tensorflow>=2.8.0
opencv-python>=4.5.0
mediapipe>=0.8.10
numpy>=1.21.0
scikit-learn>=1.0.0
tqdm>=4.62.0
gTTS>=2.3.0
pygame>=2.1.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
